SDHCI: Add support for version 3.00 and some cleanup
The SD Host Controller only has support for version 2.00.
This adds support for version 3.00 of the specification.
The main difference is that some registers now have a
slightly different meaning or size.  In particular, the
Capabilities registers have changed the way they report
supported frequencies.  In version 2.00, the frequency
divider bit field was only 8 bits and had very specific
supported values.  In version 3.00, the divider can be
any number and is 10 bits wide.  This significantly
changes how the frequency gets calculated.

BUG=chrome-os-partner:30836
TEST=Build and boot Rambi SeaBIOS

Change-Id: Ia8ad869d9568a9fd1b53f6decd3ba48a1d5d4f9c
Signed-off-by: Mike Loptien <mike.loptien@se-eng.com>
Reviewed-on: https://chromium-review.googlesource.com/228584
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>