MMC: Add function to decode EXT_CSD
The EXT_CSD contains extra information about the drive
and should be decoded like the CSD and CID fields.  This
is a 512 byte structure but many of these fields are not
important to the current MMC implementation and can be
ignored.  This only decodes some of the significant bit
fields of the EXT_CSD.

BUG=chrome-os-partner:30836
TEST=Build and boot Rambi SeaBIOS

Change-Id: Ia0cef8c89c8bb5ea012af22dbdeadb6b41475d70
Signed-off-by: Mike Loptien <mike.loptien@se-eng.com>
Reviewed-on: https://chromium-review.googlesource.com/228586
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>