SD: Move some prints to higher debug level
The SD/MMC code has a lot of debug statements in the command
generation and in the reads.  This creates a lot of extra
delay in a read command that is unnecessary.  Moving these
print statements to a higher debug level preserves them
while allowing the rest of the SD/MMC initialization code
to still print useful messages.

BUG=chrome-os-partner:30836
TEST=Build and boot Rambi SeaBIOS

Change-Id: I5ac51606e03cb2a2b78e87a03f43397eb63206d6
Signed-off-by: Mike Loptien <mike.loptien@se-eng.com>
Reviewed-on: https://chromium-review.googlesource.com/228589
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>