SD/MMC: Allow SD and MMC cards to coexist
The SD Interface structure holds all of the information
needed to setup and interface to an SD/MMC device.  This
structure was a global variable and would get filled in
after finding a compatible PCI device during a PCI scan.
However, the global variable would get overwritten by
the last device to get discovered and initialized.  So
if both an SD card and MMC drive exist in the system,
whichever one was initialized last would gain control
of the global structure and you would never be able to
boot from the other device.  By making this variable a
local, both devices are able to allocate their own SD
Interface structures and both can be booted from.

BUG=chrome-os-partner:30836
TEST=Build and boot Rambi SeaBIOS

Change-Id: I7da71fb2dc809451afea2986194049f5668b7255
Signed-off-by: Mike Loptien <mike.loptien@se-eng.com>
Reviewed-on: https://chromium-review.googlesource.com/228591
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>