SD: Fixup CMD3 to only run once
CMD3 get the RCA for SD cards.  If the command is successful
it checks the status in the response and then returns if it
is in Standby mode or tries again if not.  There is a comment
that states that the status may not be updated until the next
command is run, causing the RCA command to get executed again
in some cases.  I have found that the RCA is incorrect on this
second run and returns bad card data in future commands.  There
is a Get Status command that should be used instead to check
the status of the card and this is done outside of CMD3 so there
is no need to check the status inside CMD3.

BUG=chrome-os-partner:30836
TEST=Build and boot Rambi SeaBIOS

Change-Id: I35be6fe7fac9b7fe069402698e971e3dfac785c4
Signed-off-by: Mike Loptien <mike.loptien@se-eng.com>
Reviewed-on: https://chromium-review.googlesource.com/228592
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>