SD: Call sdhc_cmd directly instead of through pointer
The SeaBIOS build process has recently changed significantly
and there is a hole that causes sdhc_cmd to look like SeaBIOS
INIT code.  The INIT code is expected to be temporary and gets
reclaimed by the OS as it boots.  As the SD card boots to the
OS, the sdhc_cmd function gets overwritten and this causes the
board to triple fault and reset.  Calling the function directly,
instead of through the function pointer, causes the build to put
this function into code space rather than into the INIT space
and it does not get overwritten, allowing the board to boot.

BUG=chrome-os-partner:30836
TEST=Build and boot Rambi SeaBIOS

Change-Id: I857c3a1a5aa300133a4eced9a05f60ae2d86dd87
Signed-off-by: Mike Loptien <mike.loptien@se-eng.com>
Reviewed-on: https://chromium-review.googlesource.com/228593
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>