coreboot: Interpret CBFS header pointer as relative, not absolute
This patch changes the CBFS parser to interpret the CBFS header address
at the end of the CBFS image as a 32-bit signed relative offset from the
end of the image, rather than an absolute address. For the normal case
of CONFIG_CBFS_LOCATION=0 the effect is the same, but for other cases
(e.g. where SeaBIOS' own little CBFS is embedded in a larger image), it
will find the header correctly from the default cookie that cbfstool
(which always assumed to work on a full image) placed there. (This fix
has long been present in libpayload, which is unfortunately not very
useful for payloads that somehow decided not to use it.)

CQ-DEPEND=CL:229976
BRANCH=None
BUG=None
TEST=Booted with ebuild change on Falco, confirmed that SeaBIOS finds
its CBFS components (option ROM, etc.).

Change-Id: I3c28ef8f9fe0067cbd311a74cdf28655470dd05e
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/231690
Reviewed-by: Aaron Durbin <adurbin@chromium.org>